requirements.txt
anyio==4.4.0
certifi==2024.8.30
class-registry==2.1.2
contourpy==1.3.0
cycler==0.12.1
filters==1.3.2
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
kiwisolver==1.4.6
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-telegram-bot==21.5
pytz==2024.1
regex==2024.7.24
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.33
typing_extensions==4.12.2
tzdata==2024.1
